vectordb_chat/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.3.0
faiss-cpu>=1.9.0
numpy>=1.26.0
pydantic>=2.5.0
python-multipart>=0.0.6
requests
torch
python-docx
PyPDF2
